modify: tipo de atributos en el dto de reporte de carrera

+++ trabajo-integrador-tp2/src/main/java/dtos/ReporteCarreraDTO.java
@@ -3,15 +3,13 @@ package dtos;
 public class ReporteCarreraDTO {
     private String nombreCarrera;
     private int anio;
-    private long cantidadInscriptos;
-    private long cantidadEgresados;
+    private int cantidadInscriptos;
+    private int cantidadEgresados;
 
-    public ReporteCarreraDTO(String nombreCarrera, int anio, long cantidadInscriptos, long cantidadEgresados) {
+    public ReporteCarreraDTO(String nombreCarrera, int anio, int cantidadInscriptos, int cantidadEgresados) {
         this.nombreCarrera = nombreCarrera;
         this.anio = anio;
         this.cantidadInscriptos = cantidadInscriptos;
         this.cantidadEgresados = cantidadEgresados;
     }
-
-    // Getters y setters
 }
